#230 [fix] bottom sheet 파라미터 수정

+++ feature/todo/src/main/java/hous/release/feature/todo/detail/TodoDetailScreen.kt
@@ -254,8 +254,8 @@ private fun BottomSheetContent(
     getTodosAppliedFilter: () -> Unit,
     selectDayOfWeek: (Int) -> Unit,
     selectHomy: (Int) -> Unit,
-    editAction: (Int) -> Unit,
-    deleteAction: (Int) -> Unit
+    editAction: () -> Unit,
+    deleteAction: () -> Unit
 ) {
     when (selected) {
         FILTER_BOTTOM_SHEET -> {
